fix: parsing of quiqqer input fields at the settings xml generation
Changes made to the 'Panel.js' file in the '/bin/backend/controls/customer/' directory. Removed the
use of 'self' and replaced it with 'this'. Converted traditional functions to arrow functions,
which allows the 'this' keyword within the function to automatically capture its context. Improved
readability and ensured cleaner, more modern code. Changes should not affect the functionality of
the methods.