feat: add customer reference code
Added a new form field for the customer reference code in the panel. This field is defined in both
HTML and JavaScript sections of customer panel, and has been included in the locale and user
attribute XML files. This will allow users to input a reference code for customers which may assist
in customer identification or referencing.